search function and search result people cards
