+++ package.json
@@ -17,6 +17,7 @@
     "dotenv": "^16.4.7",
     "firebase": "^11.3.0",
     "firebase-admin": "^13.1.0",
+    "fuse.js": "^7.1.0",
     "next": "^15.1.6",
     "react": "^19.0.0",
     "react-dom": "^19.0.0"
