+++ src/app/administrative-staff-dashboard/page.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import AdminDashboard from "../components/AdministrativeStaffDashboard";
+
+export default function AdministrativeStaffDashboardPage() {
+    return <AdminDashboard />;
+}
