--- src/app/it-admin-dashboard/page.tsx
@@ -1,7 +1,0 @@
-"use client";
-
-import AdminDashboard from "../components/AdminDashboard";
-
-export default function ITAdminDashboardPage() {
-    return <AdminDashboard />;
-}
